+++ quantrl/backends/jax.py
@@ -6,7 +6,7 @@
 __name__    = 'quantrl.backends.jax'
 __authors__ = ["Sampreet Kalita"]
 __created__ = "2024-03-10"
-__updated__ = "2025-04-21"
+__updated__ = "2025-05-11"
 
 # dependencies
 from inspect import getfullargspec
@@ -21,7 +21,7 @@ from .base import BaseBackend
 # TODO: Implement buffers
 # TODO: Implement equinox
 
-class JaxBackend(BaseBackend):
+class JAXBackend(BaseBackend):
     """Backend to interface the JAX library.
 
     Refer to :class:`quantrl.backends.base.BaseBackend` for further documentation.
@@ -69,7 +69,7 @@ class JaxBackend(BaseBackend):
         )
 
         self.jit_repeat = jax.jit(
-            jnp.repeat,
+            lambda tensor, repeats, axis: jnp.repeat(tensor, repeats, axis),
             static_argnums=(1, 2)
         )
 
@@ -122,7 +122,7 @@ class JaxBackend(BaseBackend):
         tensor,
         dtype:str=None
     ) -> np.ndarray:
-        return np.array(tensor, dtype=self.dtype_from_str(
+        return np.asarray(tensor, dtype=self.dtype_from_str(
             dtype=dtype,
             numpy=True
         ) if dtype is not None else None)