+++ quantrl/envs/base.py
@@ -6,7 +6,7 @@
 __name__    = 'quantrl.envs.base'
 __authors__ = ["Sampreet Kalita"]
 __created__ = "2023-04-25"
-__updated__ = "2025-05-10"
+__updated__ = "2025-05-11"
 
 # dependencies
 from abc import ABC, abstractmethod
@@ -24,7 +24,6 @@ from ..backends.base import BaseBackend
 from ..io import FileIO
 from ..plotters import TrajectoryPlotter, LearningCurvePlotter
 
-# TODO: Interface ConsoleIO
 # TODO: Support for different number of states and observables
 
 class BaseEnv(ABC):
@@ -521,7 +520,7 @@ class BaseEnv(ABC):
         """
 
         # validate arguments
-        assert data_rewards is not None or n_episodes is not None, "either one of the parameters ``n_episodes`` or ``data_rewards`` should be provided"
+        assert data_rewards is not None or n_episodes is not None, "either one of the parameters ``data_rewards`` or ``n_episodes`` should be provided"
 
         # extract frequently used variables
         _idx_s = self._idx_s
@@ -941,8 +940,9 @@ class BaseGymEnv(BaseEnv, Env):
         self.io.update_cache(
             data=self.all_data if self.cache_all_data else self.data
         )
-
-        # update plot
+        # update episode reward
+        self.data_rewards.append(self.rewards)
+        # update plotter
         if self.plot and self.traj_idx % self.plot_interval == 0:
             self.plotter.plot_lines(
                 xs=self.T_norm,
@@ -953,6 +953,7 @@ class BaseGymEnv(BaseEnv, Env):
 
         # close environment
         if close:
+            self.reset()
             self.close(
                 save=False
             )
@@ -1422,7 +1423,9 @@ class BaseSB3Env(BaseEnv, VecEnv):
                 print("Batch truncated")
                 break
 
-        # update plot
+        # update episode reward
+        self.data_rewards.append(self.rewards)
+        # update plotter
         if self.plot:
             for _i in tqdm(
                 range(len(self.plotter_env_idxs)),
@@ -1439,11 +1442,9 @@ class BaseSB3Env(BaseEnv, VecEnv):
                 )
             self.plotter.hold_plot()
 
-        # update episode reward
-        self.data_rewards.append(self.rewards)
-
         # close environment
         if close:
+            self.reset()
             self.close(
                 save=save
             )