+++ quantrl/envs/deterministic.py
@@ -6,7 +6,7 @@
 __name__    = 'quantrl.envs.deterministic'
 __authors__ = ["Sampreet Kalita"]
 __created__ = "2023-04-25"
-__updated__ = "2024-10-14"
+__updated__ = "2025-05-11"
 
 # quantrl modules
 from ..backends.context_manager import get_backend_instance
@@ -57,7 +57,7 @@ class LinearizedHOEnv(BaseGymEnv):
     data_idxs: list
         Indices of the data to store into the ``data`` attribute. The indices can be selected from the complete set of values at each point of time (total ``1 + n_actions + n_observations + n_properties + 1`` elements in the same order, where the first element is the time and the last element is the reward).
     backend_library: str, default='numpy'
-        Solver to use for each step. Options are ``'torch'`` for PyTorch-based solvers, ``'jax'`` for JAX-based solvers and ``'numpy'`` for NumPy/SciPy-based solvers.
+        Solver to use for each step. Options are ``'jax'`` for JAX-based solvers, ``'torch'`` for PyTorch-based solvers and ``'numpy'`` for NumPy/SciPy-based solvers.
     backend_precision: str, default='double'
         Precision of the numerical values in the backend. Options are ``'single'`` and ``'double'``.
     backend_device: str, default='cuda'
@@ -70,7 +70,7 @@ class LinearizedHOEnv(BaseGymEnv):
         ============    ================================================
         key             value
         ============    ================================================
-        ode_method      (*str*) method used to solve the ODEs/DDEs. Available options are ``'dopri8'``, ``'dopri5'``, ``'bosh3'``, ``'fehlberg2'`` and ``'adaptive_huen'`` for a TorchDiffEq-based solver, ``'dopri8'``, ``'dopri5'`` and ``'tsit5'`` for a Diffrax-based solver and ``'BDF'``, ``'DOP853'``, ``'LSODA'``, ``'Radau'``, ``'RK23'``, ``'RK45'``, ``'dop853'``, ``'dopri5'``, ``'lsoda'``, ``'zvode'`` and ``'vode'`` for a SciPy-based solver. Default is ``'vode'``.
+        ode_method      (*str*) method used to solve the ODEs/DDEs. Available options are ``'dopri5'``, ``'dopri8'`` and ``'tsit5'`` for a Diffrax-based solver, ``'adaptive_huen'``, ``'bosh3'``, ``'dopri5'``, ``'dopri8'``, ``'fehlberg2'`` and ``'tsit5'`` for a TorchDiffEq-based solver and ``'BDF'``, ``'DOP853'``, ``'LSODA'``, ``'Radau'``, ``'RK23'``, ``'RK45'``, ``'dop853'``, ``'dopri5'``, ``'lsoda'``, ``'vode'`` and ``'zvode'`` for a SciPy-based solver. Default is ``'dopri5'``.
         ode_atol        (*float*) absolute tolerance of the ODE/DDE solver. Default is ``1e-9``.
         ode_rtol        (*float*) relative tolerance of the ODE/DDE solver. Default is ``1e-6``.
         ============    ================================================
@@ -80,13 +80,13 @@ class LinearizedHOEnv(BaseGymEnv):
     """dict: Default parameters of the environment."""
 
     default_ode_solver_params = {
-        'ode_method': 'vode',
+        'ode_method': 'dopri5',
         'ode_atol': 1e-9,
         'ode_rtol': 1e-6
     }
     """dict: Default parameters of the ODE solver."""
 
-    backend_libraries = ['torch', 'jax', 'numpy']
+    backend_libraries = ['jax', 'torch', 'numpy']
     """list: Available backend libraries."""
 
     def __init__(self,
@@ -106,14 +106,14 @@ class LinearizedHOEnv(BaseGymEnv):
         data_idxs:list,
         backend_library:str='numpy',
         backend_precision:str='double',
-        backend_device:str='gpu',
+        backend_device:str='cuda',
         dir_prefix:str='data',
         **kwargs
     ):
         """Class constructor for LinearizedHOEnv."""
 
         # validate arguments
-        assert backend_library in self.backend_libraries, f"parameter ``solver_type`` should be one of ``{self.backend_libraries}``"
+        assert backend_library in self.backend_libraries, f"parameter ``backend_library`` should be one of ``{self.backend_libraries}``"
 
         # select backend
         backend = get_backend_instance(
@@ -483,7 +483,7 @@ class LinearizedHOVecEnv(BaseSB3Env):
     data_idxs: list
         Indices of the data to store into the ``data`` attribute. The indices can be selected from the complete set of values at each point of time (total ``1 + n_actions + n_observations + n_properties + 1`` elements in the same order, where the first element is the time and the last element is the reward).
     backend_library: str, default='numpy'
-        Solver to use for each step. Options are ``'torch'`` for PyTorch-based solvers, ``'jax'`` for JAX-based solvers and ``'numpy'`` for NumPy/SciPy-based solvers.
+        Solver to use for each step. Options are  ``'jax'`` for JAX-based solvers, ``'torch'`` for PyTorch-based solvers and ``'numpy'`` for NumPy/SciPy-based solvers.
     backend_precision: str, default='double'
         Precision of the numerical values in the backend. Options are ``'single'`` and ``'double'``.
     backend_device: str, default='cuda'
@@ -496,7 +496,7 @@ class LinearizedHOVecEnv(BaseSB3Env):
         ============    ================================================
         key             value
         ============    ================================================
-        ode_method      (*str*) method used to solve the ODEs/DDEs. Available options are ``'dopri8'``, ``'dopri5'``, ``'bosh3'``, ``'fehlberg2'`` and ``'adaptive_huen'`` for a TorchDiffEq-based solver, ``'dopri8'``, ``'dopri5'`` and ``'tsit5'`` for a Diffrax-based solver and ``'BDF'``, ``'DOP853'``, ``'LSODA'``, ``'Radau'``, ``'RK23'``, ``'RK45'``, ``'dop853'``, ``'dopri5'``, ``'lsoda'``, ``'zvode'`` and ``'vode'`` for a SciPy-based solver. Default is ``'vode'``.
+        ode_method      (*str*) method used to solve the ODEs/DDEs. Available options are ``'dopri8'``, ``'dopri5'`` and ``'tsit5'`` for a Diffrax-based solver, ``'dopri8'``, ``'dopri5'``, ``'bosh3'``, ``'fehlberg2'`` and ``'adaptive_huen'`` for a TorchDiffEq-based solver and ``'BDF'``, ``'DOP853'``, ``'LSODA'``, ``'Radau'``, ``'RK23'``, ``'RK45'``, ``'dop853'``, ``'dopri5'``, ``'lsoda'``, ``'zvode'`` and ``'vode'`` for a SciPy-based solver. Default is ``'vode'``.
         ode_atol        (*float*) absolute tolerance of the ODE/DDE solver. Default is ``1e-9``.
         ode_rtol        (*float*) relative tolerance of the ODE/DDE solver. Default is ``1e-6``.
         ============    ================================================
@@ -506,13 +506,13 @@ class LinearizedHOVecEnv(BaseSB3Env):
     """dict: Default parameters of the environment."""
 
     default_ode_solver_params = {
-        'ode_method': 'vode',
+        'ode_method': 'dopri5',
         'ode_atol': 1e-9,
         'ode_rtol': 1e-6
     }
     """dict: Default parameters of the ODE solver."""
 
-    backend_libraries = ['torch', 'jax', 'numpy']
+    backend_libraries = ['jax', 'torch', 'numpy']
     """list: Available backend libraries."""
 
     def __init__(self,
@@ -533,14 +533,14 @@ class LinearizedHOVecEnv(BaseSB3Env):
         data_idxs:list,
         backend_library:str='numpy',
         backend_precision:str='double',
-        backend_device:str='gpu',
+        backend_device:str='cuda',
         dir_prefix:str='data',
         **kwargs
     ):
         """Class constructor for LinearizedHOEnv."""
 
         # validate arguments
-        assert backend_library in self.backend_libraries, f"parameter ``solver_type`` should be one of ``{self.backend_libraries}``"
+        assert backend_library in self.backend_libraries, f"parameter ``backend_library`` should be one of ``{self.backend_libraries}``"
 
         # select backend
         backend = get_backend_instance(